==> notebooks/analysis/set2/03_v2_no_merge_diversity_picking_MaxMin_2.ipynb <==
import os
import pandas as pd
from rdkit.SimDivFilters import  MaxMinPicker
from rdkit import RDLogger
RDLogger.DisableLog('rdApp.error')
from tqdm.auto import tqdm
tqdm.pandas()
from multiprocessing import Pool, cpu_count
import time
import pandas as pd
from rdkit import Chem

# load datasets pickle from ../data/ECFP_datasets/
rna_binders_df = pd.read_pickle('data_dataset2/standardized_data/ECFP_datasets_set2/rna_binders_ECFP6_set2_dedup.pkl')
rna_non_binders_df = pd.read_pickle('data_dataset2/standardized_data/ECFP_datasets_set2/rna_non_binders_ECFP6_set2_dedup.pkl')
protein_binders_df = pd.read_pickle('data_dataset2/standardized_data/ECFP_datasets_set2/protein_binders_ECFP6_set2_dedup.pkl')
protein_non_binders_df = pd.read_pickle('data_dataset2/standardized_data/ECFP_datasets_set2/protein_non_binders_ECFP6_set2_dedup.pkl')


rna_binders_df.shape, rna_non_binders_df.shape, protein_binders_df.shape, protein_non_binders_df.shape

total_molecules_rna = len(rna_binders_df) + len(rna_non_binders_df)
total_molecules_rna

# create folder for results
# create subfolder for each ECFP dataset ECFP4, ECFP6, ECFP8, ECFP10
directories = ['../data/diversity_picking/ECFP6_set2/']


for directory in directories:
    if not os.path.exists(directory):
        os.makedirs(directory)



rna_binders_df.columns

all_df = pd.concat([rna_binders_df, rna_non_binders_df, protein_binders_df, protein_non_binders_df], ignore_index=True)
all_df.shape



def check_mol_smiles_ecfp(df):
    invalid_indices = []

    for idx, row in df.iterrows():
        mol = row['mol']
        smiles = row['smiles']
        ecfp = row['ECFP6']

        # Check mol validity
        if mol is None or mol.GetNumAtoms() == 0:
            invalid_indices.append(idx)

        # Check smiles validity
        if smiles is None or smiles == '':
            invalid_indices.append(idx)
        else:
            mol_from_smiles = Chem.MolFromSmiles(smiles)
            if mol_from_smiles is None:
                invalid_indices.append(idx)

        # Check ECFP6 validity
        if ecfp is None or ecfp == '':
            invalid_indices.append(idx)

    if invalid_indices:
        invalid_rows = df[df.index.isin(invalid_indices)]
        for _, row in invalid_rows.iterrows():
            mol = row['mol']
            smiles = row['smiles']
            ecfp = row['ECFP6']

            if mol is None or mol.GetNumAtoms() == 0:
                print(f"Invalid mol at index {row.name}: {smiles}")
            if smiles is None or smiles == '' or Chem.MolFromSmiles(smiles) is None:
                print(f"Invalid smiles at index {row.name}: {smiles}")
            if ecfp is None or ecfp == '':
                print(f"Invalid ECFP6 at index {row.name}: {ecfp}")

        df.drop(invalid_indices, inplace=True)
        print("Invalid rows removed from the dataframe.")
    else:
        print("No invalid rows found.")

    return df


check_mol_smiles_ecfp(all_df)

all_df['source'].unique()

# split all_df into datasets using column 'source' an reset index
rna_binders_df = all_df[all_df['source'] == 'robin_b'].reset_index(drop=True)
rna_non_binders_df = all_df[all_df['source'] == 'robin_n'].reset_index(drop=True)
protein_binders_df = all_df[all_df['source'] == 'probes_drugs'].reset_index(drop=True)
protein_non_binders_df = all_df[all_df['source'] == 'zinc_dark_m'].reset_index(drop=True)

rna_binders_df.shape, rna_non_binders_df.shape, protein_binders_df.shape, protein_non_binders_df.shape

def diversity_picking_MaxMin(df, colname, filename):
    # Resetting the index of the DataFrame
    df = df.reset_index(drop=True)

    processed_col = df[colname]
    smiles_col = df['smiles']
    mol_col = df['mol']
    source_col = df['source']
    ecfp6_col = df['ECFP6']

    fps = processed_col.values
    mmp = MaxMinPicker()
    picks = mmp.LazyBitVectorPick(fps, len(df), len(df), seed=42)
    
    picked_df = pd.DataFrame({
        'mol': mol_col.iloc[picks],
        'smiles': smiles_col.iloc[picks],
        colname: processed_col.iloc[picks],
        'ECFP6': ecfp6_col.iloc[picks],     
        'source': source_col.iloc[picks]
    })

    picked_df.to_pickle(filename)
    return picked_df



# do diversity picking for all datasets
diversity_picking_MaxMin(rna_binders_df, 'ECFP6', '../data/diversity_picking/ECFP6_set2/rna_binders_ECFP6_set2_diversity_picked.pkl')
diversity_picking_MaxMin(rna_non_binders_df, 'ECFP6', '../data/diversity_picking/ECFP6_set2/rna_non_binders_ECFP6_set2_diversity_picked.pkl')
diversity_picking_MaxMin(protein_binders_df, 'ECFP6', '../data/diversity_picking/ECFP6_set2/protein_binders_ECFP6_set2_diversity_picked.pkl')
diversity_picking_MaxMin(protein_non_binders_df, 'ECFP6', '../data/diversity_picking/ECFP6_set2/protein_non_binders_ECFP6_set2_diversity_picked.pkl')
