==> notebooks/analysis/set2/04_v2_distance_matrix_compute.ipynb <==
import os
from rdkit import RDLogger
RDLogger.DisableLog('rdApp.error')
from rdkit import DataStructs
import numpy as np
import pandas as pd
from tqdm.auto import tqdm
tqdm.pandas()


# create directory for saving similarity matrices
if not os.path.exists('../output/distance_matrices/set2'):
    os.makedirs('../output/distance_matrices/set2')

    





rna_binders_path = '../data/diversity_picking/ECFP6_set2/rna_binders_ECFP6_set2_diversity_picked.pkl'
rna_non_binders_path = '../data/diversity_picking/ECFP6_set2/rna_non_binders_ECFP6_set2_diversity_picked.pkl'
protein_binders_path = '../data/diversity_picking/ECFP6_set2/protein_binders_ECFP6_set2_diversity_picked.pkl'
protein_non_binders_path = '../data/diversity_picking/ECFP6_set2/protein_non_binders_ECFP6_set2_diversity_picked.pkl'

rna_binders = pd.read_pickle(rna_binders_path)
n_mol = len(rna_binders)


rna_non_binders = pd.read_pickle(rna_non_binders_path).head(n_mol)
protein_binders = pd.read_pickle(protein_binders_path).head(n_mol)
protein_non_binders = pd.read_pickle(protein_non_binders_path).head(n_mol)

# Length of all dataframes
print(len(rna_binders), len(rna_non_binders), len(protein_binders), len(protein_non_binders))


# Combine the dataframes
combined_df6 = pd.concat([rna_binders, rna_non_binders, protein_binders, protein_non_binders], ignore_index=True)

n = combined_df6.shape[0]
ecfp_column = ['ECFP6']
df_list = []

for ecfp_col in tqdm(ecfp_column, desc="Calculating distance matrices"):
    if ecfp_col == 'ECFP6':
        combined_df = combined_df6

    distance_matrix = np.zeros((n, n))

    for i in range(n):
        if i % 100 == 0:
            print(i)
        for j in range(i, n):
            distance_matrix[i, j] = 1 - DataStructs.TanimotoSimilarity(combined_df[ecfp_col][i], combined_df[ecfp_col][j])
            distance_matrix[j, i] = distance_matrix[i, j]

    df = pd.DataFrame({'mol': combined_df['mol'], 'source': combined_df['source'], ecfp_col: distance_matrix.tolist()})
    df_list.append(df)

    file_name = f'../output/distance_matrices/v2/distance_matrix_{ecfp_col}_set2.pkl'
    df.to_pickle(file_name)

# Concatenate the DataFrames for all ECFP columns
combined_df = pd.concat(df_list, axis=1)

# load pickle
dist_mat_ = pd.read_pickle('../output/distance_matrices/v2/distance_matrix_ECFP6_set2.pkl')
dist_mat_

dist_mat_['source'].unique()
dist_mat_
