==> notebooks/analysis/set2/13_results_gnn.ipynb <==
import os
import pandas as pd
import os
import pandas as pd
import numpy as np
from scipy import stats
import json

# Set1 model large dataset RNA binders vs Protein binders

# df_gated = pd.read_csv('set1_ml_models/gatedgraphconv/test_results_GatedGraphConv.csv')
# df_gatv2 = pd.read_csv('set1_ml_models/gatv2conv/test_results_GATv2Conv.csv')
# df_sage = pd.read_csv('set1_ml_models/sageconv/test_results_SageConv.csv')


# Define the base directory where CSV files are located
base_dir = 'set1_ml_models'

# Models and the associated directories
models = {
    'GatedGraphConv': 'gatedgraphconv',
    'GATv2Conv': 'gatv2conv',
    'SageConv': 'sageconv'
}

# Prepare to collect all DataFrame objects
dataframes = []

# Iterate over each model and their associated directory
for model, dir_name in models.items():
    # Construct the path to the CSV file
    csv_path = os.path.join(base_dir, dir_name, f'test_results_{model}.csv')
    
    # Check if the file exists
    if os.path.exists(csv_path):
        # Load the CSV file
        df = pd.read_csv(csv_path)
        # Add a column for the model
        df['Model'] = model
        # Append the DataFrame to the list
        dataframes.append(df)
    else:
        print(f"File not found: {csv_path}")

# Concatenate all DataFrames into a single DataFrame
all_data = pd.concat(dataframes, ignore_index=True)
# rename col F1-score to F1 score
all_data.rename(columns={'F1-Score': 'F1 Score'}, inplace=True)
all_data.rename(columns={'ROC-AUC': 'ROC AUC'}, inplace=True)
# Display the combined DataFrame
all_data

# Assuming 'accuracy' is the column name that contains accuracy scores
# Find the row with the maximum accuracy
best_model_row = all_data.loc[all_data['Accuracy'].idxmax()]

# Extract information about the best model
best_accuracy = best_model_row['Accuracy']
best_model_run_id = best_model_row['Run ID']
model = best_model_row['Model']

print(f"Set1: Best {model} with the highest accuracy is from run {best_model_run_id} with an accuracy of {best_accuracy}.")


# Construct the path to the JSON file where the model's parameters are saved
best_model_dir = os.path.join(base_dir, f'sageconv/sample_{best_model_run_id:03d}')
best_model_json_path = os.path.join('set1_ml_models/sageconv/sage_best_hyperparams_run_4.json')

# Load the JSON file
with open(best_model_json_path, 'r') as file:
    model_params = json.load(file)

# Print the model parameters
print(f"Set1: Best GNN model parameters:")
for param, value in model_params.items():
    print(f"{param}: {value}")

# Models and metrics to analyze
models = ['GatedGraphConv', 'GATv2Conv', 'SageConv']
metrics = ['Accuracy', 'Precision', 'Recall', 'F1 Score', 'ROC AUC']

results = []

# Calculate metrics for each model
for model in models:
    model_data = all_data[all_data['Model'] == model]
    result_row = {'Model': model}
    for metric in metrics:
        mean = model_data[metric].mean() * 100  # Convert to percentage
        std = model_data[metric].std() * 100  # Convert to percentage
        # Format as mean ± std%
        result_row[metric] = f"{mean:.2f} ± {std:.2f}%"
    results.append(result_row)

# Convert results to DataFrame
results_df = pd.DataFrame(results)

# Identify the best scores and bold them
for metric in metrics:
    best_score = max(results_df[metric], key=lambda x: float(x.split(' ± ')[0].replace('%','')))
    results_df[metric] = results_df[metric].apply(
        lambda x: f"\\textbf{{{x}}}" if x == best_score else x
    )

# Prepare the DataFrame to LaTeX table contents without adding LaTeX table commands
# Escape LaTeX special characters
results_df.replace('%', r'\%', regex=True, inplace=True)
table_contents = results_df.to_latex(index=False, escape=False, header=False)

# Strip unwanted LaTeX commands from the table contents
table_contents = table_contents.replace("\\begin{tabular}{llllll}", "") \
                               .replace("\\toprule", "") \
                               .replace("\\midrule", "") \
                               .replace("\\bottomrule", "") \
                               .replace("\\end{tabular}", "")\
                               .replace("±", "$\pm$")

# Manually construct the LaTeX table using tabularx to fit the width of the text
latex_table = (
    "\\begin{table}[ht]\n"
    "\\centering\n"
    "\\caption{RNA Binders vs Protein Binders (Merged Dataset)}\n"
    "\\label{tab:model_performance}\n"
    "\\scriptsize\n"
    "\\begin{tabularx}{\\textwidth}{Xccccc}\n"  # Changed 'l' to 'X' for the first column
    "\\toprule\n"
    "Model & " + " & ".join(metrics) + " \\\\\n"
    "\\midrule\n"
    + table_contents.strip() +  # Remove leading/trailing whitespace
    "\\bottomrule\n"
    "\\end{tabularx}\n"
    "\\end{table}\n"
)

print(latex_table)


# Set2 model small datasets model1 RNA binders RNA nonbinders (ROBIN)

# Define the base directory where CSV files are located
base_dir = 'set2_ml_models/model1_rna_b_nb'

# Models and the associated directories
models = {
    'GatedGraphConv': 'gatedgraphconv',
    'GATv2Conv': 'gatv2conv',
    'SageConv': 'sageconv'
}

# Prepare to collect all DataFrame objects
dataframes = []

# Iterate over each model and their associated directory
for model, dir_name in models.items():
    # Construct the path to the CSV file
    csv_path = os.path.join(base_dir, dir_name, f'test_results_{model}.csv')
    
    # Check if the file exists
    if os.path.exists(csv_path):
        # Load the CSV file
        df = pd.read_csv(csv_path)
        # Add a column for the model
        df['Model'] = model
        # Append the DataFrame to the list
        dataframes.append(df)
    else:
        print(f"File not found: {csv_path}")

# Concatenate all DataFrames into a single DataFrame
all_data = pd.concat(dataframes, ignore_index=True)
# rename col F1-score to F1 score
all_data.rename(columns={'F1-Score': 'F1 Score'}, inplace=True)
all_data.rename(columns={'ROC-AUC': 'ROC AUC'}, inplace=True)
# Display the combined DataFrame
all_data

# Assuming 'accuracy' is the column name that contains accuracy scores
# Find the row with the maximum accuracy
best_model_row = all_data.loc[all_data['Accuracy'].idxmax()]

# Extract information about the best model
best_accuracy = best_model_row['Accuracy']
best_model_run_id = best_model_row['Run ID']
model = best_model_row['Model']

print(f"Set1: Best {model} with the highest accuracy is from run {best_model_run_id} with an accuracy of {best_accuracy}.")


best_model_dir = os.path.join(base_dir, f'sageconv/sample_{best_model_run_id:03d}')
best_model_json_path = os.path.join('set2_ml_models/model1_rna_b_nb/sageconv/sage_best_hyperparams_run_2.json')

# Load the JSON file
with open(best_model_json_path, 'r') as file:
    model_params = json.load(file)

# Print the model parameters
print(f"Set1: Best GNN model parameters:")
for param, value in model_params.items():
    print(f"{param}: {value}")

# Models and metrics to analyze
models = ['GatedGraphConv', 'GATv2Conv', 'SageConv']
metrics = ['Accuracy', 'Precision', 'Recall', 'F1 Score', 'ROC AUC']

results = []

# Calculate metrics for each model
for model in models:
    model_data = all_data[all_data['Model'] == model]
    result_row = {'Model': model}
    for metric in metrics:
        mean = model_data[metric].mean() * 100  # Convert to percentage
        std = model_data[metric].std() * 100  # Convert to percentage
        # Format as mean ± std%
        result_row[metric] = f"{mean:.2f} ± {std:.2f}%"
    results.append(result_row)

# Convert results to DataFrame
results_df = pd.DataFrame(results)

# Identify the best scores and bold them
for metric in metrics:
    best_score = max(results_df[metric], key=lambda x: float(x.split(' ± ')[0].replace('%','')))
    results_df[metric] = results_df[metric].apply(
        lambda x: f"\\textbf{{{x}}}" if x == best_score else x
    )

# Prepare the DataFrame to LaTeX table contents without adding LaTeX table commands
# Escape LaTeX special characters
results_df.replace('%', r'\%', regex=True, inplace=True)
table_contents = results_df.to_latex(index=False, escape=False, header=False)

# Strip unwanted LaTeX commands from the table contents
table_contents = table_contents.replace("\\begin{tabular}{llllll}", "") \
                               .replace("\\toprule", "") \
                               .replace("\\midrule", "") \
                               .replace("\\bottomrule", "") \
                               .replace("\\end{tabular}", "")\
                               .replace("±", "$\pm$")

# Manually construct the LaTeX table using tabularx to fit the width of the text
latex_table = (
    "\\begin{table}[H]\n"
    "\\centering\n"
    "\\caption{GNN models on RNA Binders vs RNA Non-Binders ($Set2\_Small$)}\n"
    "\\label{tab:model_performance2}\n"
    "\\scriptsize\n"
    "\\begin{tabularx}{\\textwidth}{Xccccc}\n"  # Changed 'l' to 'X' for the first column
    "\\toprule\n"
    "Model & " + " & ".join(metrics) + " \\\\\n"
    "\\midrule\n"
    + table_contents.strip() +  # Remove leading/trailing whitespace
    "\\bottomrule\n"
    "\\end{tabularx}\n"
    "\\end{table}\n"
)

print(latex_table)


# Set2 model small dataset model2 RNA binder (ROBIN) vs Protein Binder (Probes & Drugs)

# Define the base directory where CSV files are located
base_dir = 'set2_ml_models/model2_rna_b_prot_b'

# Models and the associated directories
models = {
    'GatedGraphConv': 'gatedgraphconv',
    'GATv2Conv': 'gatv2conv',
    'SageConv': 'sageconv'
}

# Prepare to collect all DataFrame objects
dataframes = []

# Iterate over each model and their associated directory
for model, dir_name in models.items():
    # Construct the path to the CSV file
    csv_path = os.path.join(base_dir, dir_name, f'test_results_{model}.csv')
    
    # Check if the file exists
    if os.path.exists(csv_path):
        # Load the CSV file
        df = pd.read_csv(csv_path)
        # Add a column for the model
        df['Model'] = model
        # Append the DataFrame to the list
        dataframes.append(df)
    else:
        print(f"File not found: {csv_path}")

# Concatenate all DataFrames into a single DataFrame
all_data = pd.concat(dataframes, ignore_index=True)
# rename col F1-score to F1 score
all_data.rename(columns={'F1-Score': 'F1 Score'}, inplace=True)
all_data.rename(columns={'ROC-AUC': 'ROC AUC'}, inplace=True)
# Display the combined DataFrame
all_data

# Assuming 'accuracy' is the column name that contains accuracy scores
# Find the row with the maximum accuracy
best_model_row = all_data.loc[all_data['Accuracy'].idxmax()]

# Extract information about the best model
best_accuracy = best_model_row['Accuracy']
best_model_run_id = best_model_row['Run ID']
model = best_model_row['Model']

print(f"Set1: Best {model} with the highest accuracy is from run {best_model_run_id} with an accuracy of {best_accuracy}.")


best_model_dir = os.path.join(base_dir, f'sageconv/sample_{best_model_run_id:03d}')
best_model_json_path = os.path.join('set2_ml_models/model2_rna_b_prot_b/gatedgraphconv/best_hyperparams_GatedGraphConv_10.json')

# Load the JSON file
with open(best_model_json_path, 'r') as file:
    model_params = json.load(file)

# Print the model parameters
print(f"Set1: Best GNN model parameters:")
for param, value in model_params.items():
    print(f"{param}: {value}")

# Models and metrics to analyze
models = ['GatedGraphConv', 'GATv2Conv', 'SageConv']
metrics = ['Accuracy', 'Precision', 'Recall', 'F1 Score', 'ROC AUC']

results = []

# Calculate metrics for each model
for model in models:
    model_data = all_data[all_data['Model'] == model]
    result_row = {'Model': model}
    for metric in metrics:
        mean = model_data[metric].mean() * 100  # Convert to percentage
        std = model_data[metric].std() * 100  # Convert to percentage
        # Format as mean ± std%
        result_row[metric] = f"{mean:.2f} ± {std:.2f}%"
    results.append(result_row)

# Convert results to DataFrame
results_df = pd.DataFrame(results)

# Identify the best scores and bold them
for metric in metrics:
    best_score = max(results_df[metric], key=lambda x: float(x.split(' ± ')[0].replace('%','')))
    results_df[metric] = results_df[metric].apply(
        lambda x: f"\\textbf{{{x}}}" if x == best_score else x
    )

# Prepare the DataFrame to LaTeX table contents without adding LaTeX table commands
# Escape LaTeX special characters
results_df.replace('%', r'\%', regex=True, inplace=True)
table_contents = results_df.to_latex(index=False, escape=False, header=False)

# Strip unwanted LaTeX commands from the table contents
table_contents = table_contents.replace("\\begin{tabular}{llllll}", "") \
                               .replace("\\toprule", "") \
                               .replace("\\midrule", "") \
                               .replace("\\bottomrule", "") \
                               .replace("\\end{tabular}", "")\
                               .replace("±", "$\pm$")

# Manually construct the LaTeX table using tabularx to fit the width of the text
latex_table = (
    "\\begin{table}[H]\n"
    "\\centering\n"
    "\\caption{GNN models on RNA Binders vs Protein Binders ($Set2\_Small$)}\n"
    "\\label{tab:model_performance1}\n"
    "\\scriptsize\n"
    "\\begin{tabularx}{\\textwidth}{Xccccc}\n"  # Changed 'l' to 'X' for the first column
    "\\toprule\n"
    "Model & " + " & ".join(metrics) + " \\\\\n"
    "\\midrule\n"
    + table_contents.strip() +  # Remove leading/trailing whitespace
    "\\bottomrule\n"
    "\\end{tabularx}\n"
    "\\end{table}\n"
)

print(latex_table)


# Set2 model small dataset model3 RNA Binder (ROBIN) vs Non-Binders (Merged small)

# Define the base directory where CSV files are located
base_dir = 'set2_ml_models/model3_binder_nonbinder'

# Models and the associated directories
models = {
    'GatedGraphConv': 'gatedgraphconv',
    'GATv2Conv': 'gatv2conv',
    'SageConv': 'sageconv'
}

# Prepare to collect all DataFrame objects
dataframes = []

# Iterate over each model and their associated directory
for model, dir_name in models.items():
    # Construct the path to the CSV file
    csv_path = os.path.join(base_dir, dir_name, f'test_results_{model}.csv')
    
    # Check if the file exists
    if os.path.exists(csv_path):
        # Load the CSV file
        df = pd.read_csv(csv_path)
        # Add a column for the model
        df['Model'] = model
        # Append the DataFrame to the list
        dataframes.append(df)
    else:
        print(f"File not found: {csv_path}")

# Concatenate all DataFrames into a single DataFrame
all_data = pd.concat(dataframes, ignore_index=True)
# rename col F1-score to F1 score
all_data.rename(columns={'F1-Score': 'F1 Score'}, inplace=True)
all_data.rename(columns={'ROC-AUC': 'ROC AUC'}, inplace=True)
# Display the combined DataFrame
all_data

# Assuming 'accuracy' is the column name that contains accuracy scores
# Find the row with the maximum accuracy
best_model_row = all_data.loc[all_data['Accuracy'].idxmax()]

# Extract information about the best model
best_accuracy = best_model_row['Accuracy']
best_model_run_id = best_model_row['Run ID']
model = best_model_row['Model']

print(f"Set1: Best {model} with the highest accuracy is from run {best_model_run_id} with an accuracy of {best_accuracy}.")


best_model_dir = os.path.join(base_dir, f'sageconv/sample_{best_model_run_id:03d}')
best_model_json_path = os.path.join('set2_ml_models/model2_rna_b_prot_b/sageconv/sage_best_hyperparams_run_10.json')

# Load the JSON file
with open(best_model_json_path, 'r') as file:
    model_params = json.load(file)

# Print the model parameters
print(f"Set1: Best GNN model parameters:")
for param, value in model_params.items():
    print(f"{param}: {value}")

# Models and metrics to analyze
models = ['GatedGraphConv', 'GATv2Conv', 'SageConv']
metrics = ['Accuracy', 'Precision', 'Recall', 'F1 Score', 'ROC AUC']

results = []

# Calculate metrics for each model
for model in models:
    model_data = all_data[all_data['Model'] == model]
    result_row = {'Model': model}
    for metric in metrics:
        mean = model_data[metric].mean() * 100  # Convert to percentage
        std = model_data[metric].std() * 100  # Convert to percentage
        # Format as mean ± std%
        result_row[metric] = f"{mean:.2f} ± {std:.2f}%"
    results.append(result_row)

# Convert results to DataFrame
results_df = pd.DataFrame(results)

# Identify the best scores and bold them
for metric in metrics:
    best_score = max(results_df[metric], key=lambda x: float(x.split(' ± ')[0].replace('%','')))
    results_df[metric] = results_df[metric].apply(
        lambda x: f"\\textbf{{{x}}}" if x == best_score else x
    )

# Prepare the DataFrame to LaTeX table contents without adding LaTeX table commands
# Escape LaTeX special characters
results_df.replace('%', r'\%', regex=True, inplace=True)
table_contents = results_df.to_latex(index=False, escape=False, header=False)

# Strip unwanted LaTeX commands from the table contents
table_contents = table_contents.replace("\\begin{tabular}{llllll}", "") \
                               .replace("\\toprule", "") \
                               .replace("\\midrule", "") \
                               .replace("\\bottomrule", "") \
                               .replace("\\end{tabular}", "")\
                               .replace("±", "$\pm$")

# Manually construct the LaTeX table using tabularx to fit the width of the text
latex_table = (
    "\\begin{table}[H]\n"
    "\\centering\n"
    "\\caption{GNN models on RNA-Binders vs Non-Binders ($Set2\_Small$)}\n"
    "\\label{tab:model_performance3}\n"
    "\\scriptsize\n"
    "\\begin{tabularx}{\\textwidth}{Xccccc}\n"  # Changed 'l' to 'X' for the first column
    "\\toprule\n"
    "Model & " + " & ".join(metrics) + " \\\\\n"
    "\\midrule\n"
    + table_contents.strip() +  # Remove leading/trailing whitespace
    "\\bottomrule\n"
    "\\end{tabularx}\n"
    "\\end{table}\n"
)

print(latex_table)
